# Add lock-free fastpath to callable_method_entry_or_negative