Tested with:

```
./miniruby -e '8.times.map { Ractor.new { 1_000_000.times { 1.respond_to?(:integer?) } } }.each(&:join)'"
```

Before

```
$ hyperfine "./miniruby -e '8.times.map { Ractor.new { 1_000_000.times { 1.respond_to?(:integer?) } } }.each(&:join)'"
Benchmark 1: ./miniruby -e '8.times.map { Ractor.new { 1_000_000.times { 1.respond_to?(:integer?) } } }.each(&:join)'
  Time (mean ± σ):     689.2 ms ± 275.9 ms    [User: 2165.2 ms, System: 1462.0 ms]
  Range (min … max):   480.7 ms … 1209.1 ms    10 runs
```

After

```
$ hyperfine "./miniruby -e '8.times.map { Ractor.new { 1_000_000.times { 1.respond_to?(:integer?) } } }.each(&:join)'"
Benchmark 1: ./miniruby -e '8.times.map { Ractor.new { 1_000_000.times { 1.respond_to?(:integer?) } } }.each(&:join)'
  Time (mean ± σ):     136.3 ms ±   1.3 ms    [User: 525.7 ms, System: 4.1 ms]
  Range (min … max):   133.7 ms … 139.2 ms    21 runs
```